chore(commons): extract lru-cache into commons
