+++ packages/commons/package.json
@@ -45,6 +45,10 @@
       "import": "./lib/esm/fromBase64.js",
       "require": "./lib/cjs/fromBase64.js"
     },
+    "./utils/lru-cache": {
+      "import": "./lib/esm/LRUCache.js",
+      "require": "./lib/cjs/LRUCache.js"
+    },
     "./types": {
       "import": "./lib/esm/types/index.js",
       "require": "./lib/cjs/types/index.js"
@@ -59,6 +63,10 @@
       "utils/base64": [
         "lib/cjs/fromBase64.d.ts",
         "lib/esm/fromBase64.d.ts"
+      ],
+      "utils/lru-cache": [
+        "lib/cjs/LRUCache.d.ts",
+        "lib/esm/LRUCache.d.ts"
       ],
       "types": [
         "lib/cjs/types/index.d.ts",
